requirements.txt
google-adk
crawl4ai
fastapi
uvicorn
pydantic
fastmcp
requests
PyYAML
chromadb
fastmcp

